Fix issues with the new drinjectlib process wait code
- Put back DR_EXPORT in dr_inject.h
- Fix the NT timeout code to use negative timer units
- Make the Linux impl of dr_inject_process_exit() not block
- Free drinject resources by always calling dr_inject_process_exit()

Review URL: https://codereview.appspot.com/7331051

SVN-Revision: 1917

+++ core/win32/os_exports.h
@@ -392,7 +392,7 @@ void earliest_inject_init(byte *arg_ptr);
 void earliest_inject_cleanup(byte *arg_ptr);
 
 wait_status_t
-os_wait_handle(HANDLE h, uint timeout_ms);
+os_wait_handle(HANDLE h, uint64 timeout_ms);
 
 /* in module.c */
 app_pc get_module_preferred_base_safe(app_pc pc);
